requirements.txt
# Core Framework
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart

# GraphQL
strawberry-graphql[fastapi]

# Database
motor                # Async MongoDB
pymongo             # Sync MongoDB (for Celery workers)
redis
asyncpg            # PostgreSQL async driver
sqlalchemy[asyncio]
databases
alembic

# Task Queue
celery[redis]
flower              # Celery monitoring

# Security
python-jose[cryptography]
passlib[bcrypt]
python-decouple
slowapi             # Rate limiting

# Data Processing
pandas
numpy
numba              # JIT compilation for performance
ta               # Technical indicators

# HTTP Client
httpx
aiohttp
requests

# Docker
docker

# Plotting/Visualization
matplotlib
mplfinance
plotly
kaleido

# Utilities
python-dateutil
pytz
orjson          # Fast JSON
ujson
aiofiles
uvloop             # Fast event loop

# CLI Tools
pyyaml
schedule
rich

# Development
pytest
pytest-asyncio
pytest-cov
pytest-benchmark
black
flake8
mypy
pre-commit

# Monitoring
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi

# Documentation
mkdocs
mkdocs-material

# Exchange API
python-binance

# Environment
python-dotenv

# CORS & Middleware
fastapi-cors
starlette

# WebSocket
websockets
python-socketio